LPS-55047 Split DBUpgrader so that we can apply the advice to the entire upgrade/verify method
