+++ portal-impl/src/com/liferay/portal/tools/releasecheck/ReleaseUpgrader.java
@@ -1,0 +1,177 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.releasecheck;
+
+import com.liferay.portal.dao.orm.common.SQLTransformer;
+import com.liferay.portal.events.StartupHelperUtil;
+import com.liferay.portal.kernel.cache.MultiVMPoolUtil;
+import com.liferay.portal.kernel.dao.db.DB;
+import com.liferay.portal.kernel.dao.db.DBFactoryUtil;
+import com.liferay.portal.kernel.dao.jdbc.DataAccess;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.util.ReleaseInfo;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.model.ReleaseConstants;
+import com.liferay.portal.service.ClassNameLocalServiceUtil;
+import com.liferay.portal.service.ReleaseLocalServiceUtil;
+import com.liferay.portal.service.ResourceActionLocalServiceUtil;
+import com.liferay.portal.util.PropsValues;
+import com.liferay.util.dao.orm.CustomSQLUtil;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+
+/**
+ * @author Michael C. Han
+ * @author Brian Wing Shun Chan
+ */
+public class ReleaseUpgrader extends ReleaseChecker {
+
+	public void upgrade() throws Exception {
+		// Check release
+
+		int buildNumber = ReleaseLocalServiceUtil.getBuildNumberOrCreate();
+
+		if (buildNumber > ReleaseInfo.getParentBuildNumber()) {
+			StringBundler sb = new StringBundler(6);
+
+			sb.append("Attempting to deploy an older Liferay Portal version. ");
+			sb.append("Current build version is ");
+			sb.append(buildNumber);
+			sb.append(" and attempting to deploy version ");
+			sb.append(ReleaseInfo.getParentBuildNumber());
+			sb.append(".");
+
+			throw new IllegalStateException(sb.toString());
+		}
+		else if (buildNumber < ReleaseInfo.RELEASE_5_2_3_BUILD_NUMBER) {
+			String msg = "You must first upgrade to Liferay Portal 5.2.3";
+
+			System.out.println(msg);
+
+			throw new RuntimeException(msg);
+		}
+
+		// Upgrade
+
+		if (_log.isDebugEnabled()) {
+			_log.debug("Update build " + buildNumber);
+		}
+
+		_checkPermissionAlgorithm();
+		checkReleaseState(_getReleaseState());
+
+		if (PropsValues.UPGRADE_DATABASE_TRANSACTIONS_DISABLED) {
+			disableTransactions();
+		}
+
+		try {
+			StartupHelperUtil.upgradeProcess(buildNumber);
+		}
+		catch (Exception e) {
+			updateReleaseState(ReleaseConstants.STATE_UPGRADE_FAILURE);
+
+			throw e;
+		}
+		finally {
+			if (PropsValues.UPGRADE_DATABASE_TRANSACTIONS_DISABLED) {
+				enableTransactions();
+			}
+		}
+	}
+
+	private void _checkPermissionAlgorithm() throws Exception {
+		long count = _getResourceCodesCount();
+
+		if (count == 0) {
+			return;
+		}
+
+		StringBundler sb = new StringBundler(8);
+
+		sb.append("Permission conversion to algorithm 6 has not been ");
+		sb.append("completed. Please complete the conversion prior to ");
+		sb.append("starting the portal. The conversion process is ");
+		sb.append("available in portal versions starting with ");
+		sb.append(ReleaseInfo.RELEASE_5_2_3_BUILD_NUMBER);
+		sb.append(" and prior to ");
+		sb.append(ReleaseInfo.RELEASE_6_2_0_BUILD_NUMBER);
+		sb.append(".");
+
+		throw new IllegalStateException(sb.toString());
+	}
+
+	private int _getReleaseState() throws Exception {
+		Connection con = null;
+		PreparedStatement ps = null;
+		ResultSet rs = null;
+
+		try {
+			con = DataAccess.getConnection();
+
+			ps = con.prepareStatement(
+				"select state_ from Release_ where releaseId = ?");
+
+			ps.setLong(1, ReleaseConstants.DEFAULT_ID);
+
+			rs = ps.executeQuery();
+
+			if (rs.next()) {
+				return rs.getInt("state_");
+			}
+
+			throw new IllegalArgumentException(
+				"No Release exists with the primary key " +
+					ReleaseConstants.DEFAULT_ID);
+		}
+		finally {
+			DataAccess.cleanUp(con, ps, rs);
+		}
+	}
+
+	private long _getResourceCodesCount() throws Exception {
+		Connection con = null;
+		PreparedStatement ps = null;
+		ResultSet rs = null;
+
+		try {
+			con = DataAccess.getConnection();
+
+			ps = con.prepareStatement("select count(*) from ResourceCode");
+
+			rs = ps.executeQuery();
+
+			if (rs.next()) {
+				int count = rs.getInt(1);
+
+				return count;
+			}
+
+			return 0;
+		}
+		catch (Exception e) {
+			return 0;
+		}
+		finally {
+			DataAccess.cleanUp(con, ps, rs);
+		}
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		ReleaseUpgrader.class);
+
+}
